feat: Workspace getDependencyMeta method
